Redirect hello requests to master during running state
